Struct evosim::blob::geno_blob_builder::BlobGeno

pub struct BlobGeno {
    pub vec_tree: QuadTree<GenericGenoNode>,
}

The Geno for morphyology of the blob.

The Geno is a QuadTree (it can be represented as TernaryTree as well).

index 0,1,2,3 means up,down,left,right (one of them can be ParentIndicator)

Fields
vec_tree: QuadTree<GenericGenoNode>
Implementations

impl BlobGeno

pub fn new_rand() -> BlobGeno

generate a random GenoType that don’t have conflict limbs

pub fn get_first(&self) -> Option<&GenoNode>

get the first GenoNode

pub fn is_valid(&self) -> bool

checker function to check the genotype is valid or not.

Not valid means self-conflit limbs

pub fn leaf_nodes(&self) -> Vec<usize>

all nodes don’t have child, used for mutate to lose limb

can not return root, can not return parent indicator

pub fn assign_nn_id_to_root(&mut self, id: usize)

assign an nn_id to root (sometimes builder don’t need new random geno)

pub fn all_usize_nn_ids(&self) -> Vec<usize>

get reference for all nn_id, in usize rather than Option<usize>

pub fn all_nn_ids_mut(&mut self) -> Vec<&mut Option<usize>>

get mut reference for all nn_id in the geno

pub fn all_nn_ids_indices(&self) -> Vec<usize>

pub fn move_subtree_nodes_root( &mut self, old_size: [f32; 2], new_size: [f32; 2] )

push all subtrees outside if the root block’s size changed

pub fn move_subtree_nodes( &mut self, root_index: usize, move_vec: ([f32; 2], [f32; 2], [f32; 2]) )

only used for morphyology mutation, update the center in each genonode, so that the validation check can preform

this function only works for non-root block’s mutation

pub fn change_node_size(&mut self, index: usize, new_size: [f32; 2])