Module evosim::consts::mutate_consts

Constants

MUTATE_BLOCK_SIZE_PROB

probablity of having limb size mutate

MUTATE_GAIN_LIMB_MAX_TRY

max times to retry to add a new limb if last one cause self-conflict

MUTATE_GAIN_LIMB_PROB

probablity for the choosen node to gain a child node, otherwise the blob gonna lose a limb

MUTATE_JOINT_LIMIT_MAX

MUTATE_JOINT_LIMIT_MIN

MUTATE_JOINT_LIMIT_PROB

porbablity of a signle joint limit to mutate

MUTATE_NN_PORB

porbablity of a single nn to mutate

MUTATE_NN_STD

standard deviation for normal distribution mutation

MUTATE_SINGLE_BLOCK_SIZE_CLAMP_SCALER

clamp between this scaler for DEFAULT_BLOCK_SIZE

MUTATE_SINGLE_BLOCK_SIZE_PROB

probablity for each signle block to mutate

MUTATE_SINGLE_BLOCK_SIZE_SCALER

scaler for block mutation

MUTATE_TREE_STRUCTURE_PROB

probablity of having tree structure mutate