Struct evosim::io::export::ExportFile

pub struct ExportFile {
    genovec: Vec<BlobGeno>,
    nnvec: Vec<Vec<(GenericNN, usize)>>,
    posvec: Vec<[f32; 2]>,
}

struct for file to save & load

genovec, nnvec, posvec are paired

Fields
genovec: Vec<BlobGeno>nnvec: Vec<Vec<(GenericNN, usize)>>

nested vec, outer relate to blob, inner relate to block (blob’s limb)

posvec: Vec<[f32; 2]>
Implementations

impl ExportFile

fn new() -> Self

pub fn push_blob(&mut self, blob: (&BlobGeno, &BlobInfo))

pub fn push_nn(&mut self, nnvec: Vec<(GenericNN, usize)>)

pub fn save(&self)

pub fn len(&self) -> usize

pub fn iter(&self) -> ExportFileIter<'_>

pub fn iter_mut(&mut self) -> ExportFileIterMut<'_>

pub fn check(&self)

pub fn flatten_nnvec(&self) -> Vec<GenericNN>

Flattening and sorting by usize index, return cloned nnvec